Moved fetchFieldType() method from PostRepository to the base repo (can be used by other repos as well)
