+++ src/Repositories/PostRepository.php
@@ -132,29 +132,6 @@ class PostRepository extends Repository
     }
 
     /**
-     * @param string $fieldKey
-     *
-     * @return string|null
-     */
-    public function fetchFieldType($fieldKey)
-    {
-        $post = Post::on($this->post->getConnectionName())
-                   ->orWhere(function ($query) use ($fieldKey) {
-                       $query->where('post_name', $fieldKey);
-                       $query->where('post_type', 'acf-field');
-                   })->first();
-
-        if ($post) {
-            $fieldData = unserialize($post->post_content);
-            $this->type = isset($fieldData['type']) ? $fieldData['type'] : 'text';
-
-            return $this->type;
-        }
-
-        return null;
-    }
-
-    /**
      * Get the name of the key for the field.
      *
      * @return string
